projects/kp-mezhuyev/lab2/source/requirements.txt
# Базовые зависимости
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
numpy>=1.24.0
tqdm>=4.66.0

# Для работы с текстом и токенизации
tiktoken>=0.5.1
transformers>=4.35.0

# Для embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# Для работы с markdown
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Утилиты
click>=8.1.7
pandas>=2.1.0
scikit-learn>=1.3.0

# Для векторного хранилища (этап 4)
elasticsearch>=8.0.0

# Для LLM (RAG pipeline)
openai>=1.0.0
